Web build progress
Had a lot of problems because I included SDL/SDL.h which of course used version 1.2 instead of version 2 which we link against... Didn't produce any warnings of course but resulted in crashes

+++ src/platform/sdl/sdl_handmade.cpp
@@ -236,7 +236,7 @@ DisplayBufferWindow(SDL_Renderer* renderer, const OffScreenBuffer* screenBuff, i
 
 INTERNAL void
 GetExePathAndFilename(AllState* allState) {
-    // Needed to check?
+    // TODO: Needed to check?
     const ssize_t charactersRead{ readlink("/proc/self/exe", allState->exePath.data_,
                                            allState->exePath.size) };
     allState->exeFilename = allState->exePath.data_;
@@ -250,7 +250,8 @@ GetExePathAndFilename(AllState* allState) {
 
 INTERNAL void
 BuildGamePathFilename(const AllState* allState, const char* filename, char* dest, i32 destCount) {
-    CatStrings(allState->exePath.data_, allState->exeFilename - allState->exePath.data_, filename,
+    CatStrings(allState->exePath.data_,
+               static_cast<i32>(allState->exeFilename - allState->exePath.data_), filename,
                StrLength(filename), dest, destCount);
 }
